backend/requirements.txt
eth_abi==5.1.0
eth_account==0.11.2
eth_typing==4.2.3
jsonrpcclient==4.0.3
jsonrpclib==0.2.1
jsonschema==4.22.0
py_solc_x==2.0.3
pytest==8.2.2
Requests==2.32.3
web3==6.18.0
langchain-community==0.2.5
langchain-core==0.2.7
langchain==0.2.5
openai==0.28.0
hybrid_compute_sdk==0.2.14
python-dotenv==1.0.0
six==1.16.0